% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{createObservatioPeriodHistogramTable}
\alias{createObservatioPeriodHistogramTable}
\title{TODO better by database not by ingredient}
\usage{
createObservatioPeriodHistogramTable(data, databaseId, outputFolder)
}
\description{
TODO better by database not by ingredient
}
