% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{createObservatioPeriodSummaryTable}
\alias{createObservatioPeriodSummaryTable}
\title{TODO better by database not by ingredient}
\usage{
createObservatioPeriodSummaryTable(data, databaseId, outputFolder)
}
\description{
TODO better by database not by ingredient
}
